fix datatype error in cgpa and sgpa calculation

+++ src/main/java/com/TeamEight/UniversityManagement/services/StudentService.java
@@ -8,7 +8,7 @@ public interface StudentService {
 
     Student select(String studentId);
 
-    int getCGPA(String studentId);
+    double getCGPA(String studentId);
 
-    int getSGPA(String studentId,int semester);
+    double getSGPA(String studentId,int semester);
 }
